fix(model): Fix diagram rich display in JupyterLab
JupyterLab (unlike the legacy "Jupyter notebook server") prefers JSON
over SVG for display purposes. Capellambse used to expose its internally
used JSON format in the diagrams' `_repr_mimebundle_()` method. This
caused JupyterLab to not show diagrams properly when using `display()`
or when a Diagram instance was the result of a cell.

Furthermore, the way that the `_repr_mimebundle_()` method returned its
`application/json` mimetype (encoded as a string, instead of as regular
Python dict) was never supported to begin with.

This commit fixes both of these issues by no longer exposing this
internal format publicly. This means that:

- `_repr_mimebundle_()` will no longer contain `application/json`
- `AbstractDiagram.render()` will no longer accept `json` and
  `json_pretty` as format argument
- The `AbstractDiagram.as_json` and `as_json_pretty` auto-generated
  properties are no longer present and will raise AttributeErrors