feat(cli_helpers): Recognize local directories in `loadcli`
This allows to pass the name of a local directory containing a model
into the `loadcli` family of functions and classes, and leverage the
automatic entrypoint detection logic to find a single AIRD file in it.

Also refactor the `_load_from_file` function to make it easier to read,
and remove some unnecessarily duplicated logic around JSON-in-a-string.