feat(native): Expose data to scripts via envvars
This commit introduces three new environment variables, which may be
used to simplify writing scripts that wrap a native Capella invocation:
`CAPELLA_PROJECT`, `CAPELLA_WORKSPACE` and `CAPELLA_VERSION`.

While it was already possible to figure out this information based on
how the script was called (via positional parameters and the name of the
executable), it was very inconvenient and made things for script writers
unnecessarily complicated. The new environment variables now make these
bits of information readily accessible.