fix: Don't block Capella on error in native client
When Capella errors, it creates a dialog on the display.
The Capella process is blocking until the user clicks
"Close" in the GUI, which will never happen in a CI/CD environment.

To suppress errors in the GUI, the `--launcher.suppressErrors` is
passed to Capella/Eclipse. As positive side-effect, the errors
are now also printed in stderr and are therefore handled by
capellambse.