feat(ci-templates): Allow specifying JSON as `ENTRYPOINT` everywhere
This commit changes the last remaining template to also use the
`loadcli()` function to load a model, instead of passing the
`ENTRYPOINT` variable directly to the `MelodyModel` constructor. This
allows for a lot more flexibility in how downstream users can specify
models.

The most notable improvement is that the `ENTRYPOINT` variable is now
optional in all templates for the common case where there is exactly one
`.aird` file in the repo's root. Additionally, it is now possible to
specify JSON arguments to further customize how the model is loaded.

This example performs a job on another model, which is stored in a
different subdirectory of the same repo:

```yaml
ENTRYPOINT: '{"path": "another.aird", "subdir": "different"}'
```