feat(model): Add filter and map methods to ElementList
These new methods fill two usability gaps that have existed for a while
now.

`filter` allows to specify an arbitrary function, lambda or attribute
name for filtering. The resulting list will only contains elements where
the filter result was thruthy.

`map` applies a function to each element or follows an attribute, and
builds a list with the results.

In both cases, the result will be another ElementList, meaning that
these calls can easily be chained and combined with other selectors like
the `by_X` methods.