fix(loader): Ignore specified target file in cross-fragment links
In one of our production models, we found a curious case where a
cross-fragment link, if taken literally, pointed to empty space, because
the link's target object existed in a different fragment.

Capella seems to handle this case just fine, and since we already have
everything necessary indexed anyways, we can simply check all files for
the target UUID instead of crashing.