feat(model): Enable creating new objects on non-list attributes
This is implemented with a helper class, which can be instantiated with
`capellambse.new_object()`. It takes the same arguments as the
`ElementList.create()` method. This marker object will be turned into a
real model object on assignment to an attribute.

For the moment, this is only implemented for non-list attributes made
with a DirectProxyAccessor or a RoleTagAccessor.