fix(git): Silence "file not found" logs
When trying to open a file in Git for reading which does not (yet)
exist, Git will print an error message to its stderr. This currently
gets forwarded as ERROR level log message. However, since these errors
are also turned into FileNotFoundError exceptions for proper downstream
handling, the logs are usually harmless (and, in fact, quite
misleading).

This commit changes errors that occur when trying to open a file for
reading to DEBUG level.