feat(model): Support pretty-printing text-based diagram formats
This commit introduces the new parameter `pretty_print` to the
`AbstractDiagram.render` method, which enables pretty-printing for
text-based formats like SVG. This is implemented as an optional
extension to the `DiagramFormat` protocol class, which introduces the
new method `convert_pretty` for exactly this purpose. The built-in SVG
format converter has been extended to provide this method.

This finally removes a long-standing workaround to produce
pretty-printed SVG files, which involved grabbing the underlying svg
Drawing object by rendering as the "svgdiagram" format (instead of
"svg"), and calling its methods directly to write out a pretty-printed
file - which then might have to be read back in to further work with it.
This new approach is easily extendable for any other text-based format,
and removes the need to write a temporary file.