feat(model): Add a `save` method to diagrams
This allows to quickly render a diagram and write it out to a file,
without having to juggle with file objects and (possibly) string/bytes
conversion manually.

In most cases, this works as a direct replacement for the previous
approach of rendering to a "svgdiagram" object and calling its `save`
(previously `save_diagram`) method - it also supports pretty-printing,
and it is format-agnostic. However, it does not support all possible
parameters to the svgdiagram's `save` method.

Additionally, unlike the "svgdiagram" approach, this does support
downloading and locally saving pre-rendered diagrams from the configured
diagram cache.

One problem with the current implementation is that it currently does
not support auto-detecting the format based on the file extension. We do
already have a format -> file extension association, which is currently
used to generate file names for the diagram cache. However, because
multiple formats can use the same file extension this way, it becomes
ambiguous which format we should actually write out; in my local tests,
it would consistently write `data:`-URIs into `.svg` files. For this
reason, I did not include the format auto-detection feature yet.