fix(loader): Don't fall back to making a random UUID
When a specific UUID was requested for a new element, but this ID is
already in use, raise an error instead of silently falling back to a
randomly generated ID.

This change makes some applications that may create elements more
reliable, as they no longer have to explicitly double-check whether the
ID they got is actually what they wanted - which isn't always feasible
on an application level. For example, the built-in `decl` module would
have to implement special handling for a `uuid` attribute (that may or
may not exist in the input data).