feat(http): Add a hint about 'git+' to iterdir error
It's really easy to forget adding the 'git+' prefix when copying an HTTP
clone URL from, say, Github. This will cause the URL to fall into the
HTTPFileHandler instead of the intended GitFileHandler. If the
`entrypoint` was omitted, this will raise an error that "listing files
is not supported" - which can be very confusing in this case. This
commit adds a short hint about the possibly missing protocol prefix.