feat(repl): Allow starting a REPL without a model
When specifying no model to load, the built-in REPL now starts without
loading a model, and displays a short help text on how to do so
manually. In addition, this opportunity was used to remove some rarely
used predefined launch profiles for the REPL, in favor of loading those
models manually after launch.