feat(model): Auto-generate diagram filenames
If no filename was given to a `Diagram.save()` call (i.e. the filename
argument is None), try to automatically generate an appropriate filename
based on the diagram's *name* and the format converter's filename
extension, if known. For format converters that do not specify an
extension, the name still has to be passed in explicitly.

Note that this will always place diagrams in the current working
directory of the process.