feat: Work with DRepresentationDescriptor objects
Use the DRepresentationDescriptor objects as primary interface to
diagrams. While this does add a layer of indirection, it also solves a
few issues around diagrams, allowing them to be more smoothly integrated
into the model.

`model.diagram.Diagram` now uses an `etree._Element` as its `_element`.
This allows it to be decorated as `@xtype_handler`, which enables
`ElementList` and `model.by_uuid()` to recognize them.

Due to the representation descriptors being placed in AIRD files, which
are considered not relevant for the most part, the `model.search()`
method has been extended to specifically ask the `aird` module for
representation descriptors if they're in the set of requested xtypes.

Note that this also changes the `uuid` reported by Diagram objects to
the one on the DRepresentationDescriptor instead of the DSemanticDiagram
(or similar). This matches what the Capella GUI and Java API reports,
thus making interoperability easier. However, this also means that
existing diagram caches are incompatible with the new version of
capellambse.