test: Shorten code in printed stack traces
Displaying the code inline rarely proved to actually be useful. In many
cases it just unnecessarily cluttered the test output and made it harder
to find out what is going on rather than easier. This is especially true
for parameterized tests, which tend to print the exact same code blocks
over and over again if they all fail in the same place.