perf(model): Add more efficient "contains" check
Implement the `__contains__` dunder method on ElementList with a more
efficient check. When checking whether a ModelObject is contained in the
list, we no longer have to construct proxies for comparing every
individual list member. This can speed up some client code by multiple
orders of magnitude, if a very long list or a list of lists is involved.

If the object to check for is not a model object, we fall back to
comparing proxy objects like before, since they could have overridden
the `__eq__` method. This is needed for example with EnumerationLiteral,
which compares equal to its `value` property for convenience.