feat(loader): Automate namespace management
The manual namespace management introduced in #243 was flawed: First, it
didn't support accurately versioning the introduced namespaces based on
the model's active viewpoints, and second, it had to be done manually in
several places (which it wasn't).

The new approach instead introduces a new method `update_namespaces`,
which is automatically called while saving to bring all namespace maps
up to date.