refactor(model): Adjust Enums to match metamodel
Fix the names of classes that represent `EEnum` members to match those
in the upstream metamodel. The old names are still available as
deprecated aliases. Also adjust the ordering and numbering of enum
members to match what's defined in the metamodel. And finally,
alphabetically sort the class definitions in the py file.