feat: Crop SVG diagrams to content
Background: Cases have been identified where diagrams in Capella models
have a lot of whitespace around the actual content. It was not possible
to crop the diagrams to the content using the GUI, so this feature was
implemented to automatically crop the diagrams to the content if needed.

This feature is experimental and can be enabled by setting the
environment variable
`CAPELLAMBSE_EXPERIMENTAL_CROP_SVG_DIAGRAM_CACHE_VIEWPORT` to `1`.