feat(model): Optionally enforce XHTML
This adds a new option to enforce XHTML compliance on attributes that
contain HTML markup. This feature is currently considered experimental,
and may be changed or removed in the future without warning.

This feature is useful in conjunction with templating, as it helps to
ensure that templates still contain valid (X)HTML after rendering.