feat: Implement modelling rules and validation
Note that this feature is distinct from the "Validate Model" option in
the Capella GUI. Validation in capellambse uses its own custom set of
rules.

Also note that this is only an initial version, which still has a lot of
limitations. In future iterations, the applied ruleset may be extended,
modified, or made more configurable from a user's perspective.

Co-authored-by: Ernst Würger <ernst.wuerger@deutschebahn.com>
Co-authored-by: Martin Lehmann <martin.lehmann@deutschebahn.com>
Co-authored-by: Viktor Kravchenko <viktor.kravchenko@deutschebahn.com>