fix: Prefer local over remote Docker images
When invoking a native Capella installation via Docker, `capellambse` did
always pull the image from a remote environment.

This has two disadvantages:
- It's not possible to work with local images which don't exist in a remote registry or
  if the remote registry is not reachable.
- It's vulnarable to dependency confusion since attackers could upload a vulnarable
  Docker image with the same name as the local image to DockerHub.

The new behaviour is:
`capellambse` checks if the image exists locally.
If it exists, continue with the local image. Otherwise, pull it from the remote.

This also matches the behaviour when using the `docker run` CLI.