chore: Deprecate directly hashing model objects
Keeping instances of GenericElement and its subclasses hashable has
caused problems in the past, especially in the case of custom equality
operators. Those problems include ambiguity about what gets hashed, as
well as hashing over mutable attributes, which both can cause the object
in question to end up in an unexpected hash bucket in one way or
another, which in turn leads to extremely difficult to diagnose issues
downstream. This is why hashing is now deprecated for all elements.
Previously it was only deprecated for a limited number of object types,
where it was definitely known to cause problems; this change improves
consistency across the API and prevents unexpected TypeErrors if
unhashable objects suddenly end up in hashing contexts.