feat(glart): Allow packing project/branch/job into path
This commit changes the Gitlab Artifacts (glart) file handler to parse
out the project, branch name and job name or ID from the passed `path`
argument. Additionally, the job name now defaults to the name used in
the [capella-dockerimages CI template], "update_capella_diagram_cache".

These changes make it possible in many cases to use the same URL (except
for the `glart` wrapper protocol) for both loading the model itself and
for accessing the diagram cache.

[capella-dockerimages CI template]:
  https://github.com/DSD-DBS/capella-dockerimages/blob/main/ci-templates/gitlab/diagram-cache.yml