refactor(git): Check out the work tree
This commit changes the Git file handler to check out the work tree as
real files. This simplifies some operations drastically, as it allows us
to rely more on the porcelain layer and specifically Git's index. By
having all files of the given commit checked out, it also forces all
blobs to be made available locally, and therefore resolves #427.

As part of this change, the base directory for the work trees was moved
from the system temporary directory to the user cache directory. Modern
Linux systems have the former reside in a ramdisk, which is not a good
place to store (potentially huge) model files.