feat(model): Allow mapkey without mapvalue
Specifying a mapkey but no mapvalue makes the keys map to the object
itself, instead of an attribute on the target object. This makes
str-indexing such a list equivalent to a `.by_{attr}` call with `single`
set to True:

    # assuming mapkey = "myattr", mapvalue = None
    my_list["someval"]
    # is the same as
    my_list.by_myattr("someval", single=True)

Note especially that this does not change any of the contraints around
str-indexing, specifically keys still must be unique or an exception
will be raised.