perf(aird): Use a hashmap for style lookups in iter_visible
While LXML's XPath implementation is fast, it's certainly not free.
Instead of calling an XPath once per object for the visibility check,
prebuild a dict so that lookups are constant-time.

This significantly improves performance of iter_visible (which is used
by `Diagram.nodes` in the high-level model). In a production model,
the time to compute all diagrams' visible nodes went from almost two
minutes to less than one second.