chore: Deprecate the current PVMT extension
The extension is in the process of being reworked, which will likely end
up with some existing attribute having different semantics. This affects
the `model.pvmt` and `object.pvmt` attributes, as well as the
`capellambse.extensions.pvmt` extension package.

Additionally, the `capellambse.pvmt` package is deprecated as a whole
and will be removed entirely in the future.