feat(pvmt): Rework the PVMT extension
The PVMT model extension (`capellambse.extensions.pvmt`) used to use a
parallel package to `capellambse.model` for accessing and manipulating
managed property values (`capellambse.pvmt`). This approach had several
shortcomings, which manifested in a very different API design and
missing functionality that is now commonplace in the rest of the model.

With this refactor, the PVMT model extension has been rewritten from
scratch on top of the `capellambse.model` package. This addresses the
aforementioned shortcomings and greatly improves the API, making it more
consistent and allowing to leverage new functionality.

The legacy `capellambse.pvmt` library, which had been deprecated in a
previous commit, has now been removed.

Resolves #100, resolves #104