fix(model): Strictly check RoleTagAccessor's classes
Previously when retrieving objects using a RoleTagAccessor that has a
non-empty `classes` argument, objects that didn't match the specified
classes were simply filtered out and ignored.

This caused two problems:

1. It made it more difficult to spot potential metamodel mismatches with
   upstream Capella.
2. The implementation (using `ElementList.filter`) decoupled the list
   from the model and therefore prevented modifying the model through
   it.

This commit changes the implementation to strictly verify that the
actual classes match those expected by the metamodel, and will raise an
exception in case of any mismatch.

Part of the fix to #446.