build: Add a more complete ecore parsing script
This new script is based roughly off of the old `generate_enums.py`.
However, instead of aggregating all defined classes and enums into a
single module, the new script keeps them separated by package.

Note that the script's output still needs to be adjusted manually.
However, it does help greatly in finding and covering new parts of the
metamodel, as well as keeping the definitions in sync.