chore: Update pre-commit hooks (and stage names)
The new stage names were introduced in pre-commit v3.2, and the old
names will be deprecated in v4.0.

This comes with a minor refactoring to the StackingBox. The new mypy
version understands `functools.partial` and (correctly) complained about
wrong annotations. (However it does not seem to understand a `@property`
whose setter accepts more types than the getter returns.)