revert: Don't remove unneeded namespace declarations
This commit changes the XML serialization to not remove XML namespace
declarations anymore that aren't used directly.

The implementation for that currently doesn't take into account
references to objects defined in other fragments. These references are,
on the XML level, string attributes whose content happens to reference
the namespace aliases.

In a future commit, the implementation will be updated to also consider
these cross-fragment references.