fix(model): Rework list filtering
This commit reworks the code that handles 'by_X' style list filtering.
The new code can now also filter on attributes that contain lists, where
an element matches if a filtering target is contained (or not contained)
in the list.

Note that the new implementation no longer special-cases Enum PODs for
comparison purposes, and instead delegates the string comparison to the
enum itself. This special-casing was incomplete anyways, as it wasn't
properly accounting for enums being passed in as filter argument. For
example, the following call would actually do a string comparison of the
enum member's name:

    mylist.by_kind(ComponentExchangeKind.ASSEMBLY)

This also deprecates the use of 'by_type' and 'exclude_types' for
filtering on the type of object (i.e. the Python class). These filters
collide with the commonly used 'type' attribute, which may link to
another model object of interest. Which one of these two options gets
used is very non-obvious and can lead to hard to diagnose bugs.
Instead, users are advised to use 'by_class' and 'exclude_classes', or
the '__class__' attribute for nested filter attributes.

For the time until the next major release, 'by_type' will keep the
current behavior by automatically translating it to '__class__' where
applicable. A UserWarning will be emitted in these cases to inform users
about needed actions.