feat(http): Treat more response codes as "not found"
Some servers may send a different code than 404 if a file doesn't exist
for some reason or another. For example, when requesting a PNG file from
a diagram cache that only allows SVG, the cache server may respond with
400 Bad Request instead. Other servers may respond with 410 Gone if a
file is requested that used to exist but has been deleted (because its
retention period expired, for example).