perf(diagram): Try looking up the index when loading cached diagrams
Loading a diagram from the cache may involve multiple network
round-trips, depending on the requested format and whether or not the
diagram actually exists in the cache. This can quickly become expensive.
This commit introduces some additional logic in the cache loading
routine, which uses the generated index to look up whether a cache hit
is even possible, therefore avoiding all these calls out to the network.