fix(decl): Pre-resolve 'find', 'set' and 'extend' args during sync
Before attempting to create a new object, try to pre-resolve all of the
attributes that might need to be ready at time of creation.

This fixes an issue where objects could get created multiple times, each
with only a subset of the requested operations applied to it.

This issue was already addressed in the past with 3d832fc9. However,
that commit didn't properly handle cases involving nested sync
operations.