fix(git): Cleanup the worktree from the original repo base
Previously, the 'worktree rm' command was issued from within the
worktree's GITDIR. This fails on Windows, as it cannot delete a
directory that is opened by a process (in this case implicitly as CWD).

This commit changes the finalizer to issue the 'worktree rm' command
from the base repository instead to avoid this issue.

Resolves #490