chore: Update dependencies
Update all the core and dev dependencies as well as used tool versions,
except for mypy and sphinx.

mypy has to stay on v1.15 for now, because v1.16 changes the types of
attributes that use a descriptor. This breaks our current annotations,
as they rely on the covariance of `Accessor[T]` for subclasses. mypy
v1.16 annotates those as `ElementList[T]` instead, which must be
invariant due to being a subclass of `MutableSequence[T]`. This will
have to be fixed at a later date.

Raising Sphinx to a minimum of v8.2 would make dependency resolution
impossible with uv, as capellambse still supports Python 3.10, and
therefore all its dependencies also must do so. However, Sphinx 8.2 has
dropped 3.10 support. Since we don't set an upper bound of Sphinx, this
does not prevent actually using newer versions however.