fix(exs): Deprecate encodings other than UTF-8
So far, there has been a slight difference in features between the
pure-Python and the Rust version of the XML serializer, in that the
pure-Python one accepted arbitrary encodings. However, this feature is
not exposed through the `save()` methods on MelodyLoader and
MelodyModel. Capella also uses UTF-8 for all its file read/write
operations, which makes supporting different encodings unnecessary.
Therefore this commit deprecates passing encodings into `exs`
altogether. Note that explicitly passing in "utf-8" is also included in
this deprecation, meaning that the `encoding` and `errors` arguments
will eventually be dropped completely.